RMB-862: Upgrade Vert.x from 4.1.0 to 4.1.2
Most notable fix is the "prepared statement ... already exists" issue:
https://github.com/eclipse-vertx/vertx-sql-client/pull/737

The Vertx team has merged our fix upstream therefore we no longer need
to use our fork https://github.com/folio-org/vertx-sql-client

Vert.x 4.1.2 adds three new methods to PgPool. Instead of adding
them to our multiple PgPool mocks the duplicated methods are
extracted into PgPoolBase.
