+++ doc/upgrading.md
@@ -21,26 +21,33 @@ See the [NEWS](../NEWS.md) summary of changes for each version.
 
 ## Version 33.0
 
+#### [RMB-862](https://issues.folio.org/browse/RMB-862) Upgrade to Vert.x 4.1.2
 
-#### [RMB-851](https://issues.folio.org/browse/RMB-851) Upgrade to Vert.x 4.1.0
-
-Module should depend on that version or a later version in 4.1.0 series.
+Module should depend on Vert.x 4.1.2 or a later version in 4.1.x series.
 
 Module should use `vertx-stack-depchain`:
 
 ```
-  <properties>
-    <vertx.version>4.1.0</vertx.version>
-  </properties>
   <dependencyManagement>
     <dependencies>
       <dependency>
         <groupId>io.vertx</groupId>
         <artifactId>vertx-stack-depchain</artifactId>
-        <version>${vertx.version}</version>
+        <version>4.1.2</version>
         <type>pom</type>
         <scope>import</scope>
       </dependency>
+    </dependencies>
+  </dependencyManagement>
+```
+
+The [FOLIO fork of the vertx-sql-client and vertx-pg-client](https://github.com/folio-org/vertx-sql-client/releases)
+is no longer needed because our fix has been merged upstream for
+Vert.x >= 4.1.1.
+
+Therefore _remove_ these dependencies from the pom.xml:
+
+```
       <dependency>
         <groupId>io.vertx</groupId>
         <artifactId>vertx-sql-client</artifactId>
@@ -51,8 +58,6 @@ Module should use `vertx-stack-depchain`:
         <artifactId>vertx-pg-client</artifactId>
         <version>${vertx.version}-FOLIO</version>
       </dependency>
-    </dependencies>
-  </dependencyManagement>
 ```
 
 Module should use the vertx, netty, jackson and tcnative dependencies from `vertx-stack-depchain` to avoid
